glitch fixes and cleanup
