+++ src/sprites/Portal.js
@@ -1,20 +1,12 @@
 import Phaser from 'phaser';
-import {HORIZONTAL, VERTICAL, OFFSET_PORTAL_PLACE} from '../utils/constants.js';
+import {HORIZONTAL, VERTICAL} from '../utils/constants.js';
 import {getObjectRotation, isEmpty, teleportObject, setObjectVelocityAfterPortal} from '../utils/portalUtils.js';
-import {getCollisionSide, getNearestValidWall, setPortalWallColliders} from '../utils/wallUtils.js';
+import {setPortalWallColliders} from '../utils/wallUtils.js';
 
 export default class Portal extends Phaser.Physics.Arcade.Sprite {
 	static instances = {};
 
-    constructor(scene, x, y, key, wall) {
-		//Initial setup
-		const coords = portalPlacement(x, y, wall, scene.wallLayer);
-
-		//Delete portals of the same colour other than this one
-		Portal.destroyByProperty(scene, key);
-
-		console.log(coords);
-
+    constructor(scene, coords, key) {
 		//Phaser setup (Object init)
         super(scene, coords.x, coords.y, (key + coords.orientation));
         scene.add.existing(this);
@@ -104,75 +96,5 @@ export default class Portal extends Phaser.Physics.Arcade.Sprite {
 			return false;
 		}
 		return true;
-	}
-}
-
-//Returns the x and y coordinates of where the portal should be placed or throws error if invalid
-	//Also returns orientation of the portal (horizontal or vertical)
-	//Return collision side as well; this is done here to indent the portal a bit out of the wall
-function portalPlacement(x, y, wall, wallLayer){
-	const nearestWall = getNearestValidWall(x, y, wall, wallLayer);
-	const collisionSide = getCollisionSide(x, y, wall);
-
-	if(!nearestWall){
-		throw new Error('Portal Cancel');
-	}
-
-	//Determine if connection is horizontal or vertical
-	if(wall.x == nearestWall.x){ // vertical
-		if(wall.y > nearestWall.y){
-			if(collisionSide == 'left'){
-				return {x: (nearestWall.pixelX + nearestWall.width/2 - OFFSET_PORTAL_PLACE), 
-						y: (nearestWall.pixelY + nearestWall.height),
-						orientation: VERTICAL,
-						collisionSide: 'left'};
-			}else if(collisionSide == 'right'){
-				return {x: (nearestWall.pixelX + nearestWall.width/2 + OFFSET_PORTAL_PLACE), 
-						y: (nearestWall.pixelY + nearestWall.height),
-						orientation: VERTICAL,
-						collisionSide: 'right'};
-			}
-		}else{
-			if(collisionSide == 'left'){
-				return {x: (wall.pixelX + wall.width/2 - OFFSET_PORTAL_PLACE), 
-						y: (wall.pixelY + wall.height),
-						orientation: VERTICAL,
-						collisionSide: 'left'};
-			}else if(collisionSide == 'right'){
-				return {x: (wall.pixelX + wall.width/2 + OFFSET_PORTAL_PLACE), 
-						y: (wall.pixelY + wall.height),
-						orientation: VERTICAL,
-						collisionSide: 'right'};
-			}
-			
-		}
-	}else if(wall.y == nearestWall.y){ //horizontal
-		if(wall.x > nearestWall.x){
-			if(collisionSide == 'top'){
-				return {x: (nearestWall.pixelX + nearestWall.width), 
-						y: (nearestWall.pixelY + nearestWall.height/2 - OFFSET_PORTAL_PLACE),
-						orientation: HORIZONTAL,
-						collisionSide: 'top'};
-			}else if(collisionSide == 'bottom'){
-				return {x: (nearestWall.pixelX + nearestWall.width), 
-						y: (nearestWall.pixelY + nearestWall.height/2 + OFFSET_PORTAL_PLACE),
-						orientation: HORIZONTAL,
-						collisionSide: 'bottom'};
-			}
-		}else{
-			if(collisionSide == 'top'){
-				return {x: (wall.pixelX + wall.width), 
-						y: (wall.pixelY + nearestWall.height/2 - OFFSET_PORTAL_PLACE),
-						orientation: HORIZONTAL,
-						collisionSide: 'top'};
-			}else if(collisionSide == 'bottom'){
-				return {x: (wall.pixelX + wall.width), 
-						y: (wall.pixelY + nearestWall.height/2 + OFFSET_PORTAL_PLACE),
-						orientation: HORIZONTAL,
-						collisionSide: 'bottom'};
-			}
-		}
-	}else{
-		throw new Error('Portal Cancel');
 	}
 }
